+++ CHANGELOG.md
@@ -27,6 +27,10 @@ https://github.com/wemake-services/django-modern-rest/releases/tag/0.13.0
 ### Features
 
 - Increased default `DMR_MAX_CACHE_SIZE` from `256` to `1024`, #1448
+- `slug` and `path` Django url converters now have more detailed
+  OpenAPI schemas in path parameters: `slug` has a `pattern`
+  with the regex from `SlugConverter`
+  and `path` has a `Can contain slashes` description, #1441
 
 
 ## 0.15.0 (2026-09-11)